requirements.txt
elasticsearch
langchain
python-dotenv